Remove file.
