+++ src/redux/books/books.js
@@ -1,29 +1,82 @@
-const ADD_BOOK = 'BookStores/books/ADD_BOOK';
-const REMOVE_BOOK = 'BookStores/books/REMOVE_BOOK';
+const ADD_BOOK = 'bookStore/books/ADD_BOOK';
+const REMOVE_BOOK = 'bookStore/books/REMOVE_BOOK';
+const FETCH_BOOKS = 'bookStore/books/FETCH_BOOKS';
+const DataLink = 'https://us-central1-bookstore-api-e63c8.cloudfunctions.net/bookstoreApi/apps/wOYGILJdcltu0mVjEpmf/books/';
 
 const initialState = [];
 
-// ACTION CREATORS
 export const addBook = (payload) => ({
   type: ADD_BOOK,
   payload,
 });
-export const removeBook = (bookId) => ({
-  type: REMOVE_BOOK,
-  id: bookId,
+
+export const fetchBook = (payload) => ({
+  type: FETCH_BOOKS,
+  payload,
 });
 
-// REDUCER
+export const removeBook = (payload) => ({
+  type: REMOVE_BOOK,
+  payload,
+});
 
-const reducer = (state = initialState, action) => {
+const booksReducer = (state = initialState, action) => {
   switch (action.type) {
     case ADD_BOOK:
       return [...state, action.payload];
+
     case REMOVE_BOOK:
-      return state.filter((book) => book.id !== action.id);
+      return state.filter((book) => book.id !== action.payload);
+
+    case FETCH_BOOKS:
+      return action.payload;
+
     default:
       return state;
   }
 };
 
-export default reducer;
+export const fetchBooksList = () => async (dispatch) => {
+  const booksList = await fetch(`${urlLink}`)
+    .then((response) => response.json());
+  const booksID = Object.keys(booksList);
+  const formBooks = [];
+  booksID.map((key) => formBooks.push({
+    id: key,
+    title: booksList[key][0].title,
+    author: booksList[key][0].author,
+    category: booksList[key][0].category,
+  }));
+  dispatch(fetchBook(formBooks));
+};
+
+export const postBook = (newBook) => async (dispatch) => {
+  await fetch(`${urlLink}`, {
+    method: 'POST',
+    body: JSON.stringify({
+      item_id: newBook.id,
+      title: newBook.title,
+      author: newBook.author,
+      category: newBook.category,
+    }),
+    headers: {
+      'content-type': 'application/json; charset=UTF-8',
+    },
+  });
+  dispatch(addBook(newBook));
+};
+
+export const deleteBook = (id) => async (dispatch) => {
+  await fetch(`${urlLink}/${id}`, {
+    method: 'DELETE',
+    body: JSON.stringify({
+      item_id: id,
+    }),
+    headers: {
+      'content-type': 'application/json; charset=UTF-8',
+    },
+  });
+  dispatch(removeBook(id));
+};
+
+export default booksReducer;
